Separate subgraph query runtime dependency from dev-only codegen tool
