+++ packages/queries/package.json
@@ -38,9 +38,9 @@
     "@ethersproject/transactions": "^5.6.2",
     "@ethersproject/units": "^5.6.1",
     "@snapshot-labs/snapshot.js": "^0.4.13",
-    "@synthetixio/codegen-graph-ts": "workspace:*",
     "@synthetixio/contracts": "workspace:*",
     "@synthetixio/contracts-interface": "workspace:*",
+    "@synthetixio/generate-subgraph-query": "workspace:*",
     "@synthetixio/optimism-networks": "workspace:*",
     "@synthetixio/wei": "workspace:*",
     "axios": "^0.21.4",
@@ -54,6 +54,7 @@
     "@babel/plugin-transform-runtime": "^7.13.10",
     "@babel/preset-env": "^7.13.10",
     "@babel/preset-typescript": "^7.13.0",
+    "@synthetixio/codegen-graph-ts": "workspace:*",
     "@testing-library/react-hooks": "^7.0.0",
     "@types/lodash": "^4.14.169",
     "@types/react": "^17.0.11",
