+++ tools/codegen-graph-ts/package.json
@@ -30,7 +30,6 @@
   },
   "homepage": "https://github.com/KillerByte/codegen-graph-ts#readme",
   "dependencies": {
-    "@synthetixio/wei": "workspace:*",
     "axios": "^0.21.4",
     "commander": "^8.1.0",
     "graphql": "^15.5.0",
