+++ tools/codegen-graph-ts/src/index.ts
@@ -1,6 +1,3 @@
 import { gen, pull } from './gen';
 
-import generateGql from './lib/gql';
-import axios from './lib/axios';
-
-export { generateGql, axios, gen, pull };
+export { gen, pull };
