--- tools/codegen-graph-ts/src/lib/axios.ts
@@ -1,4 +1,0 @@
-import axios from 'axios';
-
-// this is basically just a re-export which can be used by the genned code
-export default axios;
