+++ tools/generate-subgraph-query/babel.config.json
@@ -1,0 +1,4 @@
+{
+  "presets": ["@babel/preset-env", "@babel/preset-typescript"],
+  "plugins": ["@babel/transform-runtime"]
+}
