+++ tools/generate-subgraph-query/package.json
@@ -1,0 +1,52 @@
+{
+  "name": "@synthetixio/generate-subgraph-query",
+  "publishConfig": {
+    "access": "public"
+  },
+  "version": "1.0.0",
+  "description": "Generate GraphQL request to subgraph",
+  "source": "./src/index.ts",
+  "main": "./build/index.js",
+  "module": "./build/index.js",
+  "browser": "./build/index.js",
+  "types": "./build/index.d.ts",
+  "files": [
+    "build",
+    "src"
+  ],
+  "scripts": {
+    "build": "rm -rf build && tsc --build --force",
+    "lint": "eslint './src/**/*.{js,ts,tsx}' && tsc",
+    "lint:fix": "eslint --fix './src/**/*.{js,ts,tsx}'",
+    "test": "jest --coverage --no-cache"
+  },
+  "repository": "Synthetixio/js-monorepo",
+  "author": "Daniel Beal <daniel@synthetix.io>",
+  "license": "GPL-3.0",
+  "dependencies": {
+    "@synthetixio/wei": "workspace:*",
+    "lodash": "^4.17.21"
+  },
+  "devDependencies": {
+    "eslint": "^8.22.0",
+    "jest": "^27.0.6",
+    "typescript": "^4.4.2"
+  },
+  "jest": {
+    "testMatch": [
+      "<rootDir>/**/*.test.{js,jsx,ts,tsx}"
+    ],
+    "collectCoverageFrom": [
+      "<rootDir>/src/**/*.{js,jsx,ts,tsx}",
+      "!<rootDir>/src/**/*.d.{js,jsx,ts,tsx}"
+    ],
+    "coverageThreshold": {
+      "global": {
+        "branches": 0,
+        "functions": 0,
+        "lines": 0,
+        "statements": 0
+      }
+    }
+  }
+}
