+++ tools/generate-subgraph-query/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "declaration": true,
+    "declarationMap": true,
+    "module": "commonjs",
+    "outDir": "./build",
+    "rootDir": "./src"
+  },
+  "include": ["src"]
+}
